Update closureExercises.md
When looking at the source file, the numbering is correct. When viewed in the preview or viewed at https://github.com/bgando/functionalJS, the numbers after any block of JavaScript start over from 1. So problems 2 and 5 are labeled problem 1, and problems 3 and 6 are labeled problem 2. It's confusing. I'm not sure why this happens, and I don't know how to fix it. I had to change the file to propose a change, so I put in Roman numerals.